chore: move project.schema.ts to schema/ directory
